ggv2-deploy-on-device/artifacts/requirements.txt
wheel
Cython
numpy==1.19.4
flask
awsiotsdk